feat: download once

+++ Makefile
@@ -7,7 +7,7 @@ clean:
 
 download-deps:
 	mkdir -p build/deps
-	curl -o build/deps/gum.deb -L "https://github.com/charmbracelet/gum/releases/download/v0.14.0/gum_0.14.0_amd64.deb"
+	[ -f build/deps/gum.deb ] || curl -o build/deps/gum.deb -L "https://github.com/charmbracelet/gum/releases/download/v0.14.0/gum_0.14.0_amd64.deb"
 
 build: download-deps
 	time sudo ubuntu-image classic --debug tinyos.yaml
